Add float icon

+++ src/Fieldtypes/Floatval.php
@@ -6,7 +6,7 @@ use Statamic\Fields\Fieldtype;
 
 class Floatval extends Fieldtype
 {
-    protected $icon = 'integer';
+    protected $icon = 'float';
     protected $rules = ['numeric'];
     protected static $title = 'Float';
     protected static $handle = 'float';
